docs: 공지사항 관련 파일 이동
